// src/tools/colors.d.ts
export declare class ConsoleFilteredColors {
    black: string;
    red: string;
    green: string;
    yellow: string;
    blue: string;
    magenta: string;
    cyan: string;
    white: string;
    bg_black: string;
    bg_red: string;
    bg_green: string;
    bg_yellow: string;
    bg_blue: string;
    bg_magenta: string;
    bg_cyan: string;
    bg_white: string;
    reset: string;
    bright: string;
    dim: string;
    underscore: string;
    blink: string;
    reserve: string;
    hidden: string;
    strikethrough: string;
    constructor();
}
export declare class ConsoleColors extends ConsoleFilteredColors {
    black: string;
    red: string;
    green: string;
    yellow: string;
    blue: string;
    magenta: string;
    cyan: string;
    white: string;
    bg_black: string;
    bg_red: string;
    bg_green: string;
    bg_yellow: string;
    bg_blue: string;
    bg_magenta: string;
    bg_cyan: string;
    bg_white: string;
    reset: string;
    bright: string;
    dim: string;
    underscore: string;
    blink: string;
    reserve: string;
    hidden: string;
    strikethrough: string;
    constructor();
    Filtered: ConsoleFilteredColors;
}
declare const Colors: ConsoleColors;
export default Colors;